Split `update_listeners` into multiple modules
